+++ main_bot/packages/cogs/events.py
@@ -1,8 +1,11 @@
 import sys
 import re
+import traceback
+from typing import List
 
 import discord
 from discord.ext import commands
+from discord.ext.commands import Context
 
 
 class Events(commands.Cog):
@@ -28,8 +31,14 @@ class Events(commands.Cog):
             ctx = await self.client.get_context(msg)
             cmd = self.client.get_command("prefix")
             await cmd(ctx)
-       
 
+    @commands.Cog.listener()
+    async def on_command_error(self, ctx: Context, error):
+        a = sys.exc_info()
+        content = "".join(traceback.format_exception(type(error.original), error.original, error.original.__traceback__)) if \
+            hasattr(error, 'original') else "".join(traceback.format_exception(type(error), error, error.__traceback__))
+        await ctx.send(f"""```py\n{content}```""")
+       
 
 def setup(client):
     client.add_cog(Events(client))
