+++ main_bot/settings.template.py
@@ -1,0 +1,8 @@
+def bot_token():
+    return 'Bot token'
+
+def prefix():
+    return '.'
+
+def youtube_api_key():
+    return 'Youtube API Key'
